MGMT-8344: move clients package to be a typed interface
Currently clients package contains a global variable which is
a struct that contains embedded kubernetes/openshift client.
This PR defines an interface that provides all the APIs currently used.
The underlaying struct is initialized with all the needed parameters,
and the outer code is using a global Interface variable
At the later stage, the global variable will be removed and the interface
will be injected into all the packages